+++ lvs_effect_replacement_with_the_gredwitch_effects/lua/autorun/client/cl_lvs_gred_fx_override.lua
@@ -36,6 +36,7 @@ local MODULES = {
     "lvs_gred_fx/config.lua",
     "lvs_gred_fx/debug.lua",
     "lvs_gred_fx/particles.lua",
+    "lvs_gred_fx/weaponcode.lua",
     "lvs_gred_fx/muzzle.lua",
     "lvs_gred_fx/tracer.lua",
     "lvs_gred_fx/muzzleflash.lua",